pointerevents/pointerevent_pointercapture-in-shadow-dom.html

<!DOCTYPE html>
<html>
  <head>
    <title>PointerCapture for Shadow DOM Elements</title>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width">
    <link rel="help" href= "https://bugs.chromium.org/p/chromium/issues/detail?id=810882">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-actions.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script>
       class WC extends HTMLElement{
        constructor(){
          super();
          let template = document.getElementById('template-wc');
          let node = template.content.cloneNode(true) ;

          let shadowRoot = this.attachShadow({mode: 'open'});
          shadowRoot.appendChild(node);
        }
       }
       customElements.define("wc-wc", WC);
    </script>
  </head>
  <body onload="onLoad()">
    <template id="template">
      <style>
          #content{
              height:100px;
              width:100px;
              background-color: lightgrey;
          }
      </style>
      <div id="content"></div>
      <wc-wc id="wc-wc"></wc-wc>
    </template>
    <template id="template-wc">
      <style>
          #content{
              height:50px;
              width:50px;
              background-color: magenta;
          }
      </style>
      <div id="content"></div>
    </template>
    <h4>PointerCapture by Shadow DOM element</h4>
    The light gray box below is part of Shadow DOM.
    <ul>
      <li> Click left mouse button inside the box and keep mouse button depressed
      <li> Move the mouse
      <li> There should be a message stating <em>Pointer was captured by Shadow DOM!</em>
      <li> Release left mouse button
      <li> There should be a message stating <em>Pointer was released by Shadow DOM!</em>
    </ul>
    <h4>PointerCapture by Web Component Shadow DOM element</h4>
    The magenta box below is part of a Web Component's Shadow DOM.
    <ul>
      <li> Click left mouse button inside the box and keep mouse button depressed
      <li> Move the mouse
      <li> There should be a message stating <em>Pointer was captured by Web Component!</em>
      <li> Release left mouse button
      <li> There should be a message stating <em>Pointer was released by Web Component!</em>
    </ul>
    <div id="shadowhost"></div>
    <div id="log"></div>
    <script>
      var logDiv = document.getElementById("log");
      function logMessage(message){
        var log = document.createElement("div");
        var messageNode = document.createTextNode(message);
        log.appendChild(messageNode);
        logDiv.appendChild(log);
      }

      var events = [];

      var host = document.getElementById("shadowhost");
      var shadowRoot = host.attachShadow({mode: "open"});
      var template = document.getElementById('template');
      var node = template.content.cloneNode(true) ;
      shadowRoot.appendChild(node);

      var content = host.shadowRoot.getElementById("content");

      content.addEventListener("pointerdown", function(e){
        content.setPointerCapture(e.pointerId);
        events.push("pointerdown@content");
      });
      content.addEventListener("gotpointercapture", function(e){
        logMessage("Pointer was captured by Shadow DOM!");
        events.push("gotpointercapture@content");
      });
      content.addEventListener("pointerup", function(e){
        content.releasePointerCapture(e.pointerId);
        events.push("pointerup@content");
      });
      content.addEventListener("lostpointercapture", function(e){
        logMessage("Pointer was released by Shadow DOM!");
        events.push("lostpointercapture@content");
        if(window.promise_test){
          shadow_dom_test.step(function(){
            assert_array_equals(events, ["pointerdown@content",
              "gotpointercapture@content", "pointerup@content",
              "lostpointercapture@content"]);
            resolve_test();
            shadow_dom_test.done();
          });
        }
      });

      var content_wc = host.shadowRoot.getElementById("wc-wc")
         .shadowRoot.getElementById("content");

      content_wc.addEventListener("pointerdown", function(e){
        content_wc.setPointerCapture(e.pointerId);
        events.push("pointerdown@content_wc");
      });
      content_wc.addEventListener("gotpointercapture", function(e){
        logMessage("Pointer was captured by Web Component!");
        events.push("gotpointercapture@content_wc");
      });
      content_wc.addEventListener("pointerup", function(e){
        content_wc.releasePointerCapture(e.pointerId);
        events.push("pointerup@content_wc");
      });
      content_wc.addEventListener("lostpointercapture", function(e){
        logMessage("Pointer was released by Web Component!");
        events.push("lostpointercapture@content_wc");
        if(window.promise_test){
          wc_shadow_dom_test.step(function(){
            assert_array_equals(events, ["pointerdown@content_wc",
              "gotpointercapture@content_wc", "pointerup@content_wc",
              "lostpointercapture@content_wc"]);
            resolve_test();
            wc_shadow_dom_test.done();
          });
        }
      });

      var shadow_dom_test = null;
      var wc_shadow_dom_test = null;
      var resolve_test = null;
      var reject_test = null;

      function onLoad(){
        if(window.promise_test){
          promise_test(function(t){
            return new Promise(function(resolve, reject){
              shadow_dom_test = t;
              resolve_test = resolve;
              reject_test = reject;
              t.add_cleanup(function(){
                events = [];
              });
              var contentRect = content.getBoundingClientRect();
              var actions = new test_driver.Actions();
              var actions_promise = actions
                  .pointerMove(contentRect.x, contentRect.y)
                  .pointerDown({button: actions.ButtonType.LEFT})
                  .pointerUp({button: actions.ButtonType.LEFT})
                  .send();
            });
          }, "PointerCapture works for Shadow DOM element.");
          promise_test(function(t){
            return new Promise(function(resolve, reject){
              wc_shadow_dom_test = t;
              resolve_test = resolve;
              reject_test = reject;
              t.add_cleanup(function(){
                events = [];
              });
              var contentRect = content_wc.getBoundingClientRect();
              var actions = new test_driver.Actions();
              var actions_promise = actions
                  .pointerMove(contentRect.x, contentRect.y)
                  .pointerDown({button: actions.ButtonType.LEFT})
                  .pointerUp({button: actions.ButtonType.LEFT})
                  .send();
            });
          }, "PointerCapture works for Web Component Shadow DOM element.");
        }
      }
    </script>
  </body>
</html>
