LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointercapture-in-shadow-dom-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
PointerCapture by Shadow DOM element

The light gray box below is part of Shadow DOM.
Click left mouse button inside the box and keep mouse button depressed
Move the mouse
There should be a message stating Pointer was captured by Shadow DOM!
Release left mouse button
There should be a message stating Pointer was released by Shadow DOM!

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT PointerCapture works for Shadow DOM element. Test timed out

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT PointerCapture works for Shadow DOM element. Test timed out
